Module dsbuild::common::process

Definition of trait Process and struct ProcessWrapper.

Structs

Address

Process address, which is used to route network messages.

ProcessGuard

Represents read access guard for user-defined process.

ProcessWrapper

Represents wrapper around user-defined process which provides read access to it.

Enums

ProcessState

Represents possible states of the user processes inside of the system.

Traits

Process

Represents requirements for every user-defined process.