Trait dsbuild::common::process::Process

pub trait Process: Send + Sync {
    // Required methods
    fn on_local_message(
        &mut self,
        msg: Message,
        ctx: Context
    ) -> Result<(), String>;
    fn on_timer(&mut self, name: String, ctx: Context) -> Result<(), String>;
    fn on_message(
        &mut self,
        msg: Message,
        from: Address,
        ctx: Context
    ) -> Result<(), String>;
}

Represents requirements for every user-defined process.

Every user-defined process must satisfy the following requirements:

It must implement the Process trait.

Ideologically every process must be created by the user, and after that passed to the system (real or virtual), which will own the process. But technically process got static lifetime when being passed to the system, so system not holds the process. However, only system has write access to the process.

The reason for such behavior is that process can be used by user after the system is dropped, so system can not hold the process. So, for system every process’ lifetime is static.

To interact with system, process can use context Context object. It allows send messages, set timers, work with file system, etc.

Required Methods

fn on_local_message(&mut self, msg: Message, ctx: Context) -> Result<(), String>

Called when process starts interaction with system.

fn on_timer(&mut self, name: String, ctx: Context) -> Result<(), String>

Called when previously set timer is fired.

fn on_message( &mut self, msg: Message, from: Address, ctx: Context ) -> Result<(), String>

Called when process receives message.