Struct dsbuild::process_lib::ping::PingProcess

pub struct PingProcess {
    last_pong: u32,
    delay: f64,
    partner: String,
    need_cnt: u32,
    is_started: bool,
    is_stopped: bool,
}

PingProcess sends consecutive ping messages to the partner process with specified delay. It waits for the consecutive pong responses, while the last pong response is not received.

Fields
last_pong: u32

Specifies number of last received pong, initially equals to zero.

delay: f64

Specifies delay between consecutive pings.

partner: String

Specifies name of partner process, which must send pongs in response to pings. For example, it can be PongProcess.

need_cnt: u32

The number of pongs, which are needed to be received.

is_started: bool

Indicates whether the process is started.

is_stopped: bool

Indicates whether the process is stopped.

Implementations

impl PingProcess

const ON_PING_ACTIONS: usize = 2usize

Number of actions, which process will do during ping.

pub const ON_START_ACTIONS: usize = 2usize

Number of actions, which process will do during start.

pub const ON_TIMER_ACTIONS: usize = 2usize

Number of actions, which process will do during handling timer event.

pub const ON_MESSAGE_ACTIONS: usize = 0usize

Number of actions, which process will do during handling message event.

pub const ON_LAST_MESSAGE_ACTIONS: usize = 2usize

Number of actions, which process will do during handling the last message event.

pub const PING_TIMER: &'static str = "PING_TIMER"

Name of the timer, which fires if pong response do not received for a specified delay.

pub const PING_TIP: &'static str = "PING"

Name of ping message tip.

pub const PONG_TIP: &'static str = super::pong::PongProcess::PONG_TIP

Name of pong message tip.

fn ping(&self, ctx: &mut dyn Context)

Performs ping action.

Panics

In case of message can not be created, which means incorrect implementation of Message class.

pub fn get_last_pong(&self) -> u32

Returns last received pong number.

pub fn is_started(&self) -> bool

Allows to check whether the process is started.

pub fn is_stopped(&self) -> bool

Allows to check whether the process is stopped.

pub fn new(delay: f64, partner: String, need_cnt: u32) -> Self

Creates new PingProcess with delay, partner process name and need count of pongs to receive before terminate.

delay must be greater than zero.

need_cnt must be greater than zero.