Struct dsbuild::process_lib::pong::PongProcess

pub struct PongProcess {
    max_inactivity_window: f64,
    stopped: bool,
    started: bool,
    verbose: bool,
}

PongProcess waits for messages with tip equals to PING_TIP, and answers them with pong messages with tip equals to PONG_TIP and requested pong number.

Process will be stopped if there is some window of inactivity appears.

Fields
max_inactivity_window: f64

Window of inactivity after that pong process will be stopped (in seconds).

stopped: bool

Stopped flag.

started: boolverbose: bool
Implementations

impl PongProcess

const TIMER_NAME: &'static str = "PONG_TIMER"

Name of times, which fires after inactivity window.

pub const PONG_TIP: &'static str = "PONG"

Name of pong message tip.

pub const PING_TIP: &'static str = super::ping::PingProcess::PING_TIP

Name of ping message tip.

fn set_timer(&self, ctx: &mut dyn Context)

Assistant function which sets timer with specified delay and name.

pub fn new(max_inactivity_window: f64) -> Self

Creates new PongProcess with specified inactivity window (in seconds).

pub fn new_verbose(max_inactivity_window: f64) -> Self

Creates new verbose PongProcess with specified inactivity window (in seconds).

pub fn is_stopped(&self) -> bool

Check if process is stopped.

pub fn is_started(&self) -> bool

Check if process is started.