Trait dsbuild::real_mode::network::messenger::AsyncMessenger

pub trait AsyncMessenger {
    // Required methods
    fn send<'async_trait>(
        request: ProcessSendRequest
    ) -> Pin<Box<dyn Future<Output = Result<ProcessSendResponse, String>> + Send + 'async_trait>>;
    fn listen<'async_trait>(
        host: String,
        port: u16,
        pass_to: Sender<Event>
    ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>;
}

Asynchronous messenger trait, which is used to send messages between user processes.

Required Methods

fn send<'async_trait>( request: ProcessSendRequest ) -> Pin<Box<dyn Future<Output = Result<ProcessSendResponse, String>> + Send + 'async_trait>>

Creates future, which execution will lead to sending message between user processes through network, based on specified request.

Returns

Future, which execution will return Result consists of:

ProcessSendResponse with send status in case of successful sending

String will error in case of unsuccessful sending

fn listen<'async_trait>( host: String, port: u16, pass_to: Sender<Event> ) -> Pin<Box<dyn Future<Output = Result<(), String>> + Send + 'async_trait>>

Creates future, which execution will lead to sending message between user processes, through network, based on specified request.

Object Safety

This trait is not object safe.