Enum dsbuild::real_mode::real_system::AddressResolvePolicy

pub enum AddressResolvePolicy {
    Manual {
        resolve_list: Vec<Address>,
    },
}

Represents policy for resolving network address by process name.

It allows to know network address of user-process to send messages to it.

Note that address resolving does not filter out not registered processes in general1. It means user-process can receive messages from malicious processes too, which can use or not use framework;

add some kind of protection against malicious processes in future. ↩

Variants

Manual

Fields

resolve_list: Vec<Address>

List of process names and their addresses, which will be used to send messages to them.

Only processes with names correspondsaddresses in resolve_list will be resolved.

Attempts to send message to the processes which name not in resolve_list will lead to error.

All process names in the resolve_list must be unique.

Messages from not registered processes will be accepted and delivered too.

Remark: probably in future there will be possibility to add resolve records directly from user-process using context.