dsbuild
Struct Context

pub struct Context { /* private fields */ }

Handle which allows process to interact with external environment.

If implemented system is running within the simulation, then context allows to interact with simulation. Else, system is running in the real environment and context will interact with it. Using context process can send network messages to other processes, set timers, manipulate with files and sent local messages to user. Context is passed to the process by the system on every request to handle the external world event.

For more details refer to Process documentation.

Implementations

impl Context

pub fn send(&self, msg: Message, dst: Address)

Allows to unreliable send network message to the specified process.

It is not guaranteed the message will be delivered to destination.

Simulation allows to configure delivery probability and delay.

pub async fn send_with_ack( &self, msg: Message, dst: Address, timeout: f64, ) -> SendResult<()>

Allows to reliable send network message to the specified process.

On awaiting method will be blocked until the destination will receive message, send acknowledgment and it will be received by the senderโs node.

If acknowledgment was not received for timeout seconds, method will return with Timeout.

pub async fn send_with_tag( &self, msg: Message, tag: Tag, to: Address, timeout: f64, ) -> SendResult<()>

Allows to reliable send network message with specified tag.

On awaiting method will be blocked until the destination will not receive message and send acknowledgment. If acknowledgment was not received for timeout seconds, method will return with Timeout.

Note that receiver will not be explicitly notified about message was tagged, so it is on user to additionaly pass tag within the message if it is needed.

pub async fn send_recv_with_tag( &self, msg: Message, tag: Tag, to: Address, timeout: f64, ) -> SendResult<Message>

Allows to reliable send network message with specified tag and wait for response.

On awaiting method will be blocked until sender process will receive message with the same tag. On success, the received message will be returned.

If message with provided tag will not be received in timeout seconds, method will return with Timeout.

See send_with_tag documentation for more details.

pub fn spawn<F>(&self, future: F)
where F: Future<Output = ()> + Send + 'static,

Spawn asynchronous activity.

pub fn stop(self)

Allows to stop the process.

It is not guaranteed the process will be stopped immediately.

pub fn send_local(&self, message: Message)

Allows process to send message to user.

User can read process local messages in simulation using provided method.

In real mode user can read process messages using IOProcessWrapper.

pub async fn create_file<'a>(&'a self, name: &'a str) -> FsResult<File>

Allows to create file.

On success, created file will be returned.

If there is no enough space to create file, BufferSizeExceed error will be returned.

pub async fn delete_file<'a>(&'a self, name: &'a str) -> FsResult<()>

Allows to delete file.

If NotFound error is returned, file was not found. Note, that the inverse is not true. If file not exists, its removal may fail for a numbers of reasons, such as not sufficient permissions.

pub async fn file_exists<'a>(&'a self, name: &'a str) -> FsResult<bool>

Allows to check if file exists.

pub async fn open_file<'a>(&'a self, name: &'a str) -> FsResult<File>

Allows to open file with specified name.

On success, file with provided name will be returned.

If file does not exists, method will return with NotFound error.

pub fn time(&self) -> f64

Allows to get current time on node in seconds.

pub fn set_timer(&self, name: &str, delay: f64)

Set timer with specified name and delay.

If timer with such name already exists, the delay will be override.

pub fn set_timer_once(&self, name: &str, delay: f64)

Set timer with specified name and delay once.

If such timer already exists, nothing happens.

pub fn cancel_timer(&self, name: &str)

Cancel timer with specified name.