Struct dsbuild::Sim

pub struct Sim {
    inner: System,
    node_manager: Rc<RefCell<NodeManager>>,
}

Represensts simulation of real world system environment: nodes, network, time and file system.

Simulation is event-driven: in every moment there are pending events, ordered by time. Every event can cause other events and execute corresponding callback of the process-receiver. To make system to process event chain, user should call the following methods of simulation: step, make_steps, step_until_no_events.

User can add nodes to simulation with specified configuration, connect and disconnect them from network. User can add processes on some nodes and then communicate with them by sending and reading local messages. Also, user can crash nodes and recover them.

Simulation allows to configure network settings. For example, user can set delays of the network and its drop-rate.

Fields
inner: Systemnode_manager: Rc<RefCell<NodeManager>>
Implementations

impl Sim

pub fn new(seed: u64) -> Self

Create new simulation with provided seed.

pub fn set_network_delay(&self, delay: f64)

Set the fixed network delay.

pub fn set_network_delays(&self, min_delay: f64, max_delay: f64)

Set the minimum and maximum network delays.

pub fn set_network_drop_rate(&self, drop_rate: f64)

Set drop rate of the network.

pub fn connect_node_to_network(&self, node: &str)

Connect node to the network

pub fn disconnect_node_from_network(&self, node: &str)

Disconnect node from the network

pub fn split_network(&self, group1: &[&str], group2: &[&str])

Allows to disable pairwise connections between groups.

pub fn add_node(&mut self, name: &str, host: &str, port: u16)

Add node to simulation.

Node names must be unique and not contain / symbol.

Panics

In case node with such name already exists.

In case name is empty or contains / character.

pub fn add_node_with_storage( &mut self, name: &str, host: &str, port: u16, storage_capacity: usize, )

Add node with specified storage capacity to simulation.

Note that node names must be unique and not contain / symbol.

Panics

In case node with such name already exists.

In case name is empty or contains / character.

pub fn crash_node(&mut self, node_name: &str)

Crashes the specified node and its storage.

All pending events created by the node will be discarded. The undelivered messages sent by the node will be dropped. All pending and future events destined to the node will be discarded.

Processes running on the node are not cleared to allow working with processes after the crash (i.e. examine event log).

pub fn recover_node(&mut self, node_name: &str)

Recovers the previously crashed node.

Processes running on the node before the crash are cleared. The delivery of events to the node is enabled.

pub fn shutdown_node(&mut self, node_name: &str)

Shutdowns the specified node with saving storage.

pub fn rerun_node(&mut self, node_name: &str)

Reruns previously shut node.

pub fn is_node_crashed(&self, node: &str) -> bool

Checks if the node is crashed.

pub fn add_process<P: Process + 'static>( &mut self, process_name: &str, process: P, node_name: &str, ) -> ProcessWrapper<P>

Add process.

Panics

If node with such name node_name does not exists.

If process with such process name already exists on the node with name node_name.

If process name or node name is empty or contains / symbol.

pub fn process_names(&self) -> Vec<String>

Get names of all processes in the system.

pub fn read_local_messages( &mut self, proc: &str, node: &str, ) -> Option<Vec<Message>>

Extracts and returns local messages, produced by the process.

pub fn send_local_message(&mut self, proc: &str, node: &str, msg: Message)

Send local message to the process.

pub fn sent_message_count(&self, proc: &str, node: &str) -> u64

Returns the number of messages sent by the process.

pub fn received_message_count(&self, proc: &str, node: &str) -> u64

Returns the number of messages received by the process.

pub fn step_until_no_events(&mut self)

Steps through the simulation until there are no pending events left.

pub fn step_until_local_message( &mut self, proc: &str, node: &str, ) -> Result<Vec<Message>, String>

Steps through the simulation until there are no local messages.

pub fn make_steps(&mut self, steps: u32)

Perform specified number of steps through the simulation.

pub fn step(&mut self) -> bool

Perform single step through the simulation.