Struct dsbuild::VirtualSystem

pub struct VirtualSystem {
    inner: System,
    node_manager: Rc<RefCell<NodeManager>>,
}

Represents virtual system, which is responsible for interacting with user processes, time, network, and others.

System uses DSLab MP framework for simulating network, time, etc.

Fields
inner: Systemnode_manager: Rc<RefCell<NodeManager>>
Implementations

impl System

pub fn new(seed: u64) -> Self

Create new System with provided seed.

pub fn network(&self) -> RefMut<'_, Network>

Returns a mutable reference to network.

pub fn add_node(&mut self, name: &str, host: &str, port: u16)

Adds a node to the DSLabSimulation. Note that node names must be unique and does not contain / symbol.

Panics

In case node with such name already exists.

In case name is empty or contains / character.

pub fn crash_node(&mut self, node_name: &str)

Crashes the specified node.

All pending events created by the node will be discarded. The undelivered messages sent by the node will be dropped. All pending and future events destined to the node will be discarded.

Processes running on the node are not cleared to allow working with processes after the crash (i.e. examine event log).

pub fn recover_node(&mut self, node_name: &str)

Recovers the previously crashed node.

Processes running on the node before the crash are cleared. The delivery of events to the node is enabled.

fn get_mut_node(&self, name: &str) -> Option<RefMut<'_, Node>>

Returns a mutable reference to the node.

Can not make method public because process names on dslab nodes are not the same as in the framework.

pub fn is_node_crashed(&self, node: &str) -> bool

Checks if the node is crashed.

pub fn add_process<P: Process + Clone + 'static>( &mut self, process_name: &str, process: P, node_name: &str ) -> ProcessWrapper<P>

Adds a process to the [VirtualSystem].

Panics

If node with such name node_name does not exists.

If process with such process name is already exists on the node with node_name.

If process name or node name is empty or contains / symbol.

pub fn process_names(&self) -> Vec<String>

Returns the names of all processes in the system.

pub fn read_local_messages(&mut self, proc: &str, node: &str) -> Vec<Message>

Extracts and returns local messages, produced by the process.

pub fn send_local_message(&mut self, proc: &str, node: &str, msg: Message)

Send local message to the process.

pub fn sent_message_count(&self, proc: &str, node: &str) -> u64

Returns the number of messages sent by the process.

pub fn received_message_count(&self, proc: &str, node: &str) -> u64

Returns the number of messages received by the process.

pub fn step_until_no_events(&mut self)

Steps through the DSLabSimulation until there are no pending events left.

pub fn make_steps(&mut self, steps: u32)

Perform steps steps through the DSLabSimulation.

pub fn step(&mut self) -> bool

Perform single step through the DSLabSimulation.