dsbuild
Trait Process

pub trait Process: Send + Sync {
    // Required methods
    fn on_local_message(&mut self, msg: Message, ctx: Context);
    fn on_timer(&mut self, name: String, ctx: Context);
    fn on_message(&mut self, msg: Message, from: Address, ctx: Context);
}

Represents requirements for every user-defined process.

When process receives local message from user, timer is fired or network message is received, the corresponding process callback will be called.

To interact with system, process can use passed Context object, which represents proxy between process and external environment.

Required Methods

fn on_local_message(&mut self, msg: Message, ctx: Context)

Called when process receives local message from user.

See documentation of IOProcessWrapper struct for real mode and corresponding method of simulation for more details.

fn on_timer(&mut self, name: String, ctx: Context)

Called when previously set timer is fired.

See corresponding method of context for more details.

fn on_message(&mut self, msg: Message, from: Address, ctx: Context)

Called when process receives network message from other process.