Struct dsbuild::virt::node::NodeManager

pub struct NodeManager {
    name_to_address: HashMap<String, Address>,
    address_to_name: HashMap<Address, String>,
    node_processes: HashMap<String, HashSet<String>>,
}

Represents node manager.

WARNING: Node manager does not permit nodes and processes with names, contains /. Allowing this can potentially lead to problems with mapping to and from DSLab MP process names in form of node_name/process_name.

Fields
name_to_address: HashMap<String, Address>address_to_name: HashMap<Address, String>node_processes: HashMap<String, HashSet<String>>
Implementations

impl NodeManager

pub fn add_node( &mut self, name: String, host: String, port: u16 ) -> Result<(), String>

Add new node to the manager.

Returns

Error in case of node with such name or with such host and port already exists.

Error in case of node name is empty or contains / character.

Ok in any other case.

pub fn add_process_to_node( &mut self, node_name: String, process_name: String ) -> Result<Address, String>

Add new process to the node.

Returns

Error in case of node with such node_name does not exists.

Error in case of process with such process_name already exists on the node.

Error in case of node_name is empty or contains / character.

Error in case of process_name is empty or contains / character.

Ok with address of the added process in any other case.

pub fn clear_node(&mut self, node_name: &str)

Removed all processes from the node.

pub fn get_full_process_name(&self, address: &Address) -> Result<String, String>

Map process by it’s address into the view node_name/process_name, in which process can be added to the dslab_mp).

Returns

Error in case node with such address is not exists.

Error in case process with such name is not exists on the node.

Ok with mapped process name in case of success.

pub fn get_process_address(&self, process_name: &str) -> Result<Address, String>

Map full process name, potentially received from NodeManager::get_full_process_name(), to the process address.

Returns

Error in case process_name is not in the view of node_name/process_name.

Error if the node with such node_name does not exists.

Error if the process with such process_name does not exists on the node.

Ok with mapped process address in case of success.

pub fn check_node_exists(&self, node_name: &str) -> bool

Check if node with such node_name exists.

pub fn check_process_exists(&self, node_name: &str, process_name: &str) -> bool

Check if process with such process_name exists on the node with such node_name.

pub fn construct_full_process_name( &self, process_name: &str, node_name: &str ) -> Result<String, String>

Returns full name of the process process_name located on the node node_name.

Returns

Error in case process with such process_name does not exists on the node with such node_name.

Ok with full process name in case of success.

fn check_name(name: &str) -> Result<(), String>

Check if the name is valid.

Returns

Error in case of name is empty or contains / character.

Ok in any other case.

Such behavior is required to allow determinate mapping from node_name/proc_name view into process address.