Struct dsbuild::virtual_mode::process_wrapper::VirtualProcessWrapper

pub struct VirtualProcessWrapper<P: Process + Clone + 'static> {
    process_address: Address,
    user_process: Arc<RwLock<P>>,
    process_state: ProcessState,
    node_manager: Rc<RefCell<NodeManager>>,
}

Represents virtual process wrapper, which is to be passed to the DSLab MP.

Fields
process_address: Addressuser_process: Arc<RwLock<P>>process_state: ProcessStatenode_manager: Rc<RefCell<NodeManager>>
Implementations

impl<P: Process + Clone + 'static> VirtualProcessWrapper<P>

pub fn new( process_address: Address, process_impl: Arc<RwLock<P>>, node_manager: Rc<RefCell<NodeManager>> ) -> Self

Create new virtual process wrapper.

fn create_context(&self) -> VirtualContext

Create virtual context, which matches to the virtual process wrapper.

fn handle_process_actions( &mut self, actions: Vec<ProcessAction>, ctx: &mut SimulationContext )

Handle process actions and transform them to the DSLab MP actions.