Crate dsbuild

DSbuild is a high-level framework aimed to provide foundation for building distributed systems with support for simulation-based testing. The framework guarantees system behaviour will be consistent in both virtual and real modes, which can significantly simplify testing and debugging.

Basic concepts

DSbuild supports building aribitary distributed systems which are composed of user-defined processes. Processes can interact with network, file system, time and user. Using framework, the implemented systems can run in the real word environment on the single work station (real mode) or on multiple virtual nodes within the simulation.

Real mode. Using real mode, system can run in the real world environment with ability to communicate in network, use file system and time. The primitives provided by DSBuild guarantees system behaviour consistency in real mode and simulation.

Simulation. The main purpose of simulation is to make process of testing distributed systems easy. Network, time and file system are mocked in the simulation. It allows user to get fine-grained control over them and set it’s parameters according to the particular system specification. User can create few virtual nodes and run the implemented distributed system on them. Framework guarantees that behaviour of system in simulation will be the same as in real mode with respect to user settings of virtual environment. This approach allows to test and debug system in very complicated scenarios which can help user to find deep bugs and significantly increase system reliability. Also, as the sources of system random behaviour are deterministically mocked in simulation, it guarantees reproducibility of results from launch to launch.

Node. Node hosts processes and manages network, file system and time. In simulation there can be several nodes communicating in the network. User can disconnect it from network or crash. In the real mode there is only one node corresponding to user’s work station. It both modes processes on the same node share access to network, file system and time.

Network. In simulation, network represents abstraction other the nodes’ communication environment. User can control delay and drop rate of network. Also, simulation allows to make network partitions or disconnect nodes from it.

File system. File system represents abstraction over node’s storage. In real mode it allows to manipulate with files in the specified dirrectory. In virtual mode file system is simulated and allows user to configure bandwith and other settings of the node storage.

Process. Processes implemented by user define behaviour of the system. In particular, user can specify response of process on previously set timer fired event or when process receives message from other process via network. Process can interact with the outside world (simulation or real environment). It particular, process can exchange messages with other processes using network, request read or write to the node’s file system, set timers and send local messages to user. DSBuild allows processes to use both callbacks and asynchronous programming approach to cope with the tasks listed above.

Structs

Address

Process address, which is used to route network messages.

Context

Handle which allows process to interact with external environment.

File

Abstraction over the file.

IOProcessWrapper

Wrapper around user-defined process, which allows to send and receive local messages from it in real mode.

Message

Represents message, which is used by processes to communicate with each other by the network.

ProcessGuard

Represents read access guard for user-defined process.

ProcessWrapper

Represents wrapper around user-defined process which provides read access to it.

RealNode

Represents node in real mode.

Sim

Represensts simulation of real world system environment: nodes, network, time and file system.

Enums

FsError

Represents error types which can appear when process interacts with file system.

SendError

Represents error type of send operation.

Traits

Process

Represents requirements for every user-defined process.

Type Aliases

FsResult

Represents result of the file system operation.

SendResult

Represents result of send operation.

Tag

Represents message tag.