Crate dsbuild

Framework for building distributed systems with support for DSLab MP simulation-based testing.

Modules

common 🔒

Definition of structures and functions, which are used by real and virtual systems.

examples

Collection of framework usage examples.

process_lib

Collection of distributed process implementations, which are used in framework tests.

real_mode 🔒

Definition of structures and functions, which are used in real mode.

virtual_mode 🔒

Definition of structures and functions, which are used in virtual mode.

Structs

Address

Represents process address, which is used in real system to route network messages.

Message

Represents message, which can be used by processes to communicate with each other.

ProcessGuard

Represents guard for user-defined process.

ProcessWrapper

Represents wrapper around user-defined process, which returns to user when he passes process to real or virtual system.

RealSystem

Represents real system, which is responsible for interacting with user-processes, time, network, and other OS features.

RealSystemConfig

Represents configuration of RealSystem.

VirtualSystem

Represents virtual system, which is responsible for interacting with user-processes, simulating time, network, and other OS features.

Enums

AddressResolvePolicy

Represents policy for resolving network address by process name.

Traits

Context

Represents proxy, which provides process-system interaction.

Process

Represents requirements for every user-defined process.