Crate dsbuild

Framework for building distributed systems with support for DSLab MP simulation-based testing.

Modules

common 🔒

Definition of structures and functions, which are used by real and virtual systems.

real 🔒

storage

Definition of storage errors.

virt 🔒

Definition of structures and functions, which are used in virtual mode.

Structs

Address

Represents process address, which is used in real system and virtual system to route network messages.

Context

Represents proxy, which provides process-system interaction.

IOProcessWrapper

Represents process wrapper around user-defined process, which allows to send and receive local messages from the process.

Message

Represents message, which can be used by processes to communicate with each other.

ProcessGuard

Represents guard for user-defined process.

ProcessWrapper

Represents wrapper around user-defined process, which returns to user when he passes process to real or virtual system.

RealSystem

Represents real system.

VirtualSystem

Represents virtual system, which is responsible for interacting with user processes, time, network, and others.

Traits

Process

Represents requirements for every user-defined process.

Type Aliases

Tag

Represents a message tag, which can be used to await particular messages.